% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputChecks.R
\name{checkNumericMatrix}
\alias{checkNumericMatrix}
\title{Description of inputChecks function}
\usage{
checkNumericMatrix(x, varName)
}
\value{
true or stop
}
\description{
These function are for checking user input
}
\keyword{internal}

