% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\title{Description of the coef function for zeroSum s3 objects}
\usage{
coef(fit = NULL, s = "lambda.min", precision = 1e-06, ...)
}
\arguments{
\item{fit}{fit object generated by zeroSumCVFit() or zeroSumFit()}

\item{s}{determines which lambda of a zeroSumCVFit should be returned:
lambda.min or lamda.1SE}

\item{precision}{all coefficients below this thresshold are set to 0}

\item{...}{other arguments for the normal predict function if the fit
is not a zeroSumFit or zeroSumCVFit object}
}
\value{
estimated coefficients
}
\description{
This function returns the coefficients of a zeroSum fit object.
}
\examples{
set.seed(1)
x <- log2(exampleData$x+1)
y <- exampleData$y
fit <- zeroSumCVFit( x, y, alpha=1)
coef(fit, s="lambda.min")

}
