% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.zeroSum}
\alias{coef.zeroSum}
\title{Description of the coef function for zeroSum s3 objects}
\usage{
\method{coef}{zeroSum}(object = NULL, s = "lambda.min", ...)
}
\arguments{
\item{object}{fit object generated by zeroSum()}

\item{s}{determines which lambda of a zeroSumCVFit should be returned:
lambda.min or lamda.1SE}

\item{...}{other arguments for the normal predict function if the fit
is not a zeroSumFit object}
}
\value{
estimated coefficients
}
\description{
This function returns the coefficients of a zeroSum object object.
}
\examples{
set.seed(1)
x <- log2(exampleData$x + 1)
y <- exampleData$y
fit <- zeroSum(x, y, alpha = 1)
coef(fit, s = "lambda.min")
}
