% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvStopCheck.R
\name{cvStopCheck}
\alias{cvStopCheck}
\title{Description of devianceStop function}
\usage{
cvStopCheck(cvError, maxWorse)
}
\value{
true or false
}
\description{
This function checks if the CV-Log-Likelihood got worse in the last steps
}
\keyword{internal}

