% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/devianceStop.R
\name{devianceStop}
\alias{devianceStop}
\title{Description of devianceStop function}
\usage{
devianceStop(devRatio, maxWorse)
}
\value{
true or false
}
\description{
This function checks if the dev ratio got worse in the last steps
}
\keyword{internal}

