% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{example  data}
\format{
a \code{RangedData} instance, rows=samples.
}
\source{
simulated
}
\usage{
exampleData
}
\value{
example data
}
\description{
This data is used for demonstration
purpose / debugging
}
\author{
simulated
}
