% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRegressionDataToCSVs.R
\name{exportRegressionDataToCSV}
\alias{exportRegressionDataToCSV}
\title{Description of exportRegressionDataToCSV function
exports all zeroSum settings for using the c version of zeroSum}
\usage{
exportRegressionDataToCSV(x, y, lambda = 0, lambdaSteps = 100, alpha = 1,
  weights = NULL, penalty.factor = NULL, standardize = TRUE,
  epsilon = NULL, nFold = 10, foldid = NULL, useOffset = TRUE,
  verbose = FALSE, type = "gaussianZS", precision = 1e-08,
  diagonalMoves = TRUE, polish = 0, cvStop = 0.1, path = "",
  name = "")
}
\description{
Description of exportRegressionDataToCSV function
exports all zeroSum settings for using the c version of zeroSum
}
\keyword{internal}
