% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToCSV.R
\name{exportToCSV}
\alias{exportToCSV}
\title{Description of exportToCSV function
exports all zeroSum settings for using the c version of zeroSum}
\usage{
exportToCSV(
  x,
  y,
  family = "gaussian",
  alpha = 1,
  lambda = NULL,
  lambdaSteps = 100,
  weights = NULL,
  penalty.factor = NULL,
  zeroSum.weights = NULL,
  nFold = NULL,
  foldid = NULL,
  epsilon = NULL,
  standardize = FALSE,
  intercept = TRUE,
  zeroSum = TRUE,
  threads = "auto",
  cvStop = 0.1,
  path = "",
  name = "",
  ...
)
}
\description{
Description of exportToCSV function
exports all zeroSum settings for using the c version of zeroSum
}
\keyword{internal}
