% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNullModels.R
\name{getLogisticNullModel}
\alias{getLogisticNullModel}
\title{Description of getNullModel function}
\usage{
getLogisticNullModel(y, worg)
}
\value{
different terms and finale cost of the objective function
}
\description{
This function calculates a null model for given y,w
}
\keyword{internal}
