% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromCSVs.R
\name{importFromCSV}
\alias{importFromCSV}
\title{Description of importFromCSV function
imports zeroSum cvfit csv from the c version of zeroSum}
\usage{
importFromCSV(rDataObj, cOutput)
}
\description{
Description of importFromCSV function
imports zeroSum cvfit csv from the c version of zeroSum
}
\keyword{internal}
