% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRegressionDataFromCSVs.R
\name{importRegressionDataFromCSV}
\alias{importRegressionDataFromCSV}
\title{Description of importRegressionDataFromCSV function
imports zeroSum cvfit csv from the c version of zeroSum}
\usage{
importRegressionDataFromCSV(rDataObj, cOutput)
}
\description{
Description of importRegressionDataFromCSV function
imports zeroSum cvfit csv from the c version of zeroSum
}
\keyword{internal}
