% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.zeroSum}
\alias{plot.zeroSum}
\title{Description of the plot function for zeroSum s3 objects}
\usage{
\method{plot}{zeroSum}(x = NULL, main = "", gamma = 0, ...)
}
\arguments{
\item{x}{fit object generated by zeroSum()}

\item{main}{title of the generated plot}

\item{gamma}{if a fusion CV was performed the only a cv plot for
a selected gamma can be shown}

\item{...}{other arguments for the normal predict function if
the fit is not a zeroSumFit object}
}
\value{
predicted values
}
\description{
This function plots the CV-Error and its standard deviation
against the value of lambda of a zeroSumCVFit.
}
\examples{
set.seed(1)
x <- log2(exampleData$x + 1)
y <- exampleData$y
fit <- zeroSum(x, y, alpha = 1)
plot(fit, "test")

}
