% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.zeroSum}
\alias{predict.zeroSum}
\title{Description of the predict function for zeroSum s3 objects}
\usage{
\method{predict}{zeroSum}(object = NULL, newx = NULL, s = "lambda.min", type = NULL, ...)
}
\arguments{
\item{object}{fit object generated by zeroSum()}

\item{newx}{data matrix for which new responses should be predicted
(rows==samples)}

\item{s}{determines which lambda should be used for predicting: lambda.min or
lamda.1SE.}

\item{type}{determines the return type: "link" returns the linear predictor
(data %*% coefs), "response" returns the probabilites if applied on a
binomial or multinomial zeroSum fit. For the other cases "response"
equals "link". The type "class" returns the predicted class label if
applied on a binomial or multinomial zeroSum fit. For cox regeression
the linear predictor is returned as default ("link") and
exp(linear predictor)) for the type "response".}

\item{...}{other arguments for the normal predict function if the fit
is not a zeroSum object}
}
\value{
predicted values
}
\description{
This function uses zeroSumFit objects to predict responses for
given data newx.
}
\examples{
set.seed(1)
x <- log2(exampleData$x + 1)
y <- exampleData$y
fit <- zeroSum(x, y, alpha = 1)
predict(fit, x, s = "lambda.min")
}
