% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Description of the simulateData function}
\usage{
simulateData(coefs = c(rnorm(10), rep(0, 50)), corBetas = NULL,
  samples = 20, responseSD = 0.1, sampleWiseSD = 1)
}
\arguments{
\item{coefs}{true coeffients used for simulating a data set}

\item{corBetas}{correlation Matrix for the supplied beta Matrix}

\item{samples}{number of samples}

\item{responseSD}{standard deviation of random noise added
       to the response}

\item{sampleWiseSD}{standard deviation of the sample-wise shifts}
}
\value{
simulated data set
}
\description{
This function simulates a data set from a given linear model.
}
\examples{
data <- simulateData()
dim( data$x )
length( data$y )
}

