% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorElNetCostFunction.R
\name{vectorElNetCostFunction}
\alias{vectorElNetCostFunction}
\title{Description of vectorElNetCostFunction function}
\usage{
vectorElNetCostFunction(x, y, beta, lambda, alpha)
}
\arguments{
\item{x}{data as numeric matrix object (rows=samples)}

\item{y}{response vector (length(y)==nrow(x))}

\item{beta}{coefficients of the linear model}

\item{lambda}{penalizing parameter of the elastic-net regularization}

\item{alpha}{Lasso/Ridge adjustment: For alpha = 0 the elastic net becomes
             a ridge regularization, for alpha = 1 the elastic net becomes
             the lasso regularization}
}
\value{
value of cost function
}
\description{
This function calculates the cost function for the elastic net regression.
}
\keyword{internal}

