% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zeroSumAnalyticSolution.R
\name{zeroSumAnalyticSolution}
\alias{zeroSumAnalyticSolution}
\title{Description of zeroSumAnalyticSolution function}
\usage{
zeroSumAnalyticSolution(x, y, offset = TRUE)
}
\arguments{
\item{x}{data as a numeric matrix object (rows=samples).
         The zero-sum regression requires data on the log scale, i.e.
         x should be log-transformed data.}

\item{y}{response vector to be predicted by x (length(y)==nrow(x))}

\item{offset}{determines if an offset should be used in the model or
              not (TRUE/FALSE)}
}
\value{
zeroSumFitObject
}
\description{
Solves the unpenalized least-squares problem for given x and y under the
zero-sum constraint. However this can only be used if p<n, where p=ncol(x)
and n=length(y).
}
\examples{
set.seed(1)
data <- simulateData( coefs = rnorm(21), samples = 20)
zeroSumAnalyticSolution( data$x, data$y)
}

