% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zeroSumCVFit_subsetFit.R
\name{zeroSumCVFit_subsetFit}
\alias{zeroSumCVFit_subsetFit}
\title{Description of zeroSumCVFit_subsetFit function}
\usage{
zeroSumCVFit_subsetFit(beta, x, y, val, lambda, alpha, offset, type, algorithm,
  precision, diagonalMoves, polish)
}
\arguments{
\item{beta}{coefficients used as warm start for optimization}

\item{x}{data as a numeric matrix object (rows=samples). For zero-sum
         this x has to be log-transformed data!}

\item{y}{response vector to be predicted by x (length(y)==nrow(x))}

\item{val}{samples to be left out}

\item{lambda}{penalizing parameter of the elastic-net regularization}

\item{alpha}{Lasso/Ridge adjustment: For alpha = 0 the elastic net becomes
             a ridge regularization, for alpha = 1 the elastic net becomes
             the lasso regularization}

\item{offset}{determines if an offset should be used in the
              model or not (TRUE/FALSE)}

\item{type}{choose the regression type: elNet, zeroSumElNet}

\item{algorithm}{determines the used algorithm:
           CD = Coordinate descent (very fast, not so accurate),
           CD+LS = Coordinate descent + local search (fast, very accurate),
           LS = Local search (slow, accurate),
           SA = Simulated annealing (very slow, very accurate)}

\item{precision}{stopping criterion of the used algorithms. Determines how
                 small the improvement of the cost function has to be to stop
                 the algorithm.}

\item{diagonalMoves}{allows the coordinate descent to use diagonal moves}

\item{polish}{enables a local search at the end of CD to polish the result}
}
\value{
cross validated Fit
}
\description{
Creates an elastic net cross validated regression for x,y given alpha.
}
\keyword{internal}

