% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSumCVFitObject.R
\name{zeroSumCVFitObject}
\alias{zeroSumCVFitObject}
\title{Description of zeroSumCVFitObject function}
\usage{
zeroSumCVFitObject(obj)
}
\value{
zeroSumCVFitObject
}
\description{
Creates a zeroSumCVFitObject which stores all arguments and results of
the zeroSumCVFit() function.
}
\keyword{internal}
