% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSumFit.R
\name{zeroSumFit}
\alias{zeroSumFit}
\title{Description of zeroSumFit function}
\usage{
zeroSumFit(x, y, lambda, alpha, offset = TRUE, type = "zeroSumElNet",
  algorithm = "CD+LS", precision = 1e-06, diagonalMoves = TRUE,
  polish = 0, verbose = FALSE, beta = NULL)
}
\arguments{
\item{x}{data as a numeric matrix object (rows=samples). 
The zero-sum regression requires data on the log scale, i.e.
x should be log-transformed data.}

\item{y}{response vector to be predicted by x (length(y)==nrow(x))}

\item{lambda}{penalizing parameter of the elastic-net regularization}

\item{alpha}{Lasso/Ridge adjustment: For alpha = 0 the elastic net becomes
a ridge regularization, for alpha = 1 the elastic net becomes
the lasso regularization}

\item{offset}{determines if an offset should be used in the
model or not (TRUE/FALSE)}

\item{type}{choose the regression type: elNet, zeroSumElNet}

\item{algorithm}{choose an algorithm:
CD = Coordinate descent (very fast, not so accurate),
CD+LS = Coordinate descent + local search (fast, very accurate),
LS = Local search (slow, accurate),
SA = Simulated annealing (very slow, very accurate)}

\item{precision}{stopping criterion of the used algorithms. Determines how
small the improvement of the cost function has to be to stop
the algorithm. Default is 1e-6.}

\item{diagonalMoves}{allows the coordinate descent to use diagonal moves}

\item{polish}{enables a local search at the end of CD to polish the result}

\item{verbose}{verbose = TRUE enables output}

\item{beta}{start coeffiencts of algorithm}
}
\value{
zeroSumFitObject
}
\description{
Creates a linear model with the elastic-net regularization and the zero-sum
     constraint for given x, y, lambda and alpha.
}
\examples{
set.seed(1)
x <- log2(exampleData$x+1)
y <- exampleData$y
zeroSumFit( x, y, 1.5, 1)

}

