% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSumFitObject.R
\name{zeroSumFitObject}
\alias{zeroSumFitObject}
\title{Description of zeroSumFitObject function}
\usage{
zeroSumFitObject(lambda, alpha, coef, type, algorithm, diagonalMoves)
}
\value{
zeroSumFitObject
}
\description{
Creates a zeroSumFitObject which stores all arguments and results of
the zeroSumFit() function.
}
\keyword{internal}

