% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSumObject.R
\name{zeroSumObject}
\alias{zeroSumObject}
\title{Description of zeroSumObject function}
\usage{
zeroSumObject(obj)
}
\value{
zeroSumCVFitObject
}
\description{
Creates a zeroSumCVFitObject which stores all arguments and results of
the zeroSumCVFit() function.
}
\keyword{internal}
