% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSumRegression.R
\name{zeroSumRegression}
\alias{zeroSumRegression}
\title{Description of zeroSumRegression_C Wrapper function}
\usage{
zeroSumRegression(data, CV)
}
\value{
regression coefficients are returned
}
\description{
This is a wrapper function for the C regression functions
}
\keyword{internal}
