% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputChecks.R
\docType{data}
\name{zeroSumTypes}
\alias{zeroSumTypes}
\title{Description of inputChecks function}
\format{
An object of class \code{data.frame} with 4 rows and 2 columns.
}
\usage{
zeroSumTypes
}
\value{
true or stop
}
\description{
These function are for checking user input
}
\keyword{internal}
