% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extCostFunction.R
\name{extCostFunction}
\alias{extCostFunction}
\title{Description of costFunction function
wraps input into a regression object and
calls the costFunction and returns the result}
\usage{
extCostFunction(x, y, beta, lambda, alpha, gamma = 0, cSum = 0,
  type = zeroSumTypes[1, 1], weights = NULL, zeroSumWeights = NULL,
  penalty.factor = NULL, fusion = NULL, standardize = FALSE)
}
\description{
Description of costFunction function
wraps input into a regression object and
calls the costFunction and returns the result
}
\keyword{internal}
