% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogisticNullModel.R
\name{getLogisticNullModel}
\alias{getLogisticNullModel}
\title{Description of getLogisticNullModel function}
\usage{
getLogisticNullModel(y, worg, iterations)
}
\value{
different terms and finale cost of the objective function
}
\description{
This function calculates a null model for given x,y
}
\keyword{internal}
