% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Description of the plot function for zeroSum s3 objects}
\usage{
plot(fit = NULL, main = "", ...)
}
\arguments{
\item{fit}{fit object generated by zeroSumCVFit()}

\item{main}{title of the generated plot}

\item{...}{other arguments for the normal predict function if
             the fit is not a zeroSumCVFit object}
}
\value{
predicted values
}
\description{
This function plots the CV-Error and its standard deviation
against the value of lambda of a zeroSumCVFit.
}
\examples{
set.seed(1)
data <- simulateData()
fit <- zeroSumCVFit( data$x, data$y, alpha=1)
plot( fit, "test")
}

