% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionObject.R
\name{regressionObject}
\alias{regressionObject}
\title{Description of regresion object}
\usage{
regressionObject(x, y, type, alpha, lambda, lambdaSteps, weights,
  penalty.factor, zeroSum.weights, nFold, foldid, epsilon, standardize,
  intercept, useZeroSum, threads, cvStop, ...)
}
\description{
Description of regresion object
}
\keyword{internal}
