% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zeroSumCVFitObject.R
\name{zeroSumCVFitObject}
\alias{zeroSumCVFitObject}
\title{Description of zeroSumCVFitObject function}
\usage{
zeroSumCVFitObject(lambdaSeq, CVError, CVError_SD, alpha, devRatio, type,
  algorithmCV, algorithmAllSamples, lambdaMin, betaMin, lambda1SE, beta1SE,
  offset, precision, verticalMoves, polish, numberOfBetas)
}
\value{
zeroSumCVFitObject
}
\description{
Creates a zeroSumCVFitObject which stores all arguments and results of
the zeroSumCVFit() function.
}
\keyword{internal}

