% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSumFit.R
\name{zeroSumFit}
\alias{zeroSumFit}
\title{Description of zeroSumFit function}
\usage{
zeroSumFit(x, y, lambda, alpha = 1, weights = NULL, penalty.factor = NULL,
  standardize = FALSE, useOffset = TRUE, type = "gaussianZS",
  precision = 1e-08, diagonalMoves = FALSE, polish = TRUE,
  verbose = FALSE, beta = NULL, ...)
}
\arguments{
\item{x}{data as a numeric matrix object (rows=samples).
The zero-sum regression requires data on the log scale, i.e.
x should be log-transformed data.}

\item{y}{response vector to be predicted by x (length(y)==nrow(x))}

\item{lambda}{penalizing parameter of the elastic-net regularization}

\item{alpha}{Lasso/Ridge adjustment: For alpha = 0 the elastic net becomes
a ridge regularization, for alpha = 1 the elastic net becomes
the lasso regularization}

\item{weights}{sample weights (must be greater than zero)}

\item{penalty.factor}{weights for the elatic net regularization
(must be greater than or equal to zero)}

\item{standardize}{standardize the data (be careful standardization causes that
the scale of the data affects the coefficients! This can act contray to
scale invariance caused by the zero-sum constraint!)}

\item{useOffset}{determines if an offset should be used in the
model or not (TRUE/FALSE)}

\item{type}{choose the regression type:
\describe{
        \item{gaussian:}{}
        \item{gaussianZS:}{}
        \item{binomial:}{}
        \item{binomialZS:}{}
        \item{multinomial:}{}
        \item{multinomialZS:}{}
}}

\item{precision}{stopping criterion of the used algorithms. Determines how
small the improvement of the cost function has to be to stop
the algorithm. Default is 1e-6.}

\item{diagonalMoves}{allows the CD to use diagonal moves (can in rare cases
slightly increase the accuracy of the models but
increases the computing time)}

\item{polish}{enables a local search at the end of CD to polish the result
(removes small numeric uncertainties causing very small but non-zero
coeffiencts and has only minimal effects on the computing time)}

\item{verbose}{verbose = TRUE enables additional output about the regression}

\item{beta}{start coeffiencts of the algorithm}

\item{...}{can be used for adjusting internal parameters}
}
\value{
zeroSumFitObject
}
\description{
Creates a linear model with the elastic-net regularization and the zero-sum
     constraint for given x, y, lambda and alpha.
}
\examples{
set.seed(1)
x <- log2(exampleData$x+1)
y <- exampleData$y
zeroSumFit( x, y, lambda=1.5 )

}
