% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroSumFit.R
\name{zeroSumFit}
\alias{zeroSumFit}
\title{Description of zeroSumFit function}
\usage{
zeroSumFit(x, y, lambda, alpha = 1, weights = NULL, penalty.factor = NULL,
  zeroSumWeights = NULL, cSum = 0, standardize = FALSE, gamma = 0,
  fusion = NULL, useOffset = TRUE, useApprox = TRUE, downScaler = 1,
  type = "gaussianZS", algorithm = "CD", precision = 1e-08,
  diagonalMoves = TRUE, polish = 0, verbose = FALSE, beta = NULL)
}
\arguments{
\item{x}{data as a numeric matrix object (rows=samples).
The zero-sum regression requires data on the log scale, i.e.
x should be log-transformed data.}

\item{y}{response vector to be predicted by x (length(y)==nrow(x))}

\item{lambda}{penalizing parameter of the elastic-net regularization}

\item{alpha}{Lasso/Ridge adjustment: For alpha = 0 the elastic net becomes
a ridge regularization, for alpha = 1 the elastic net becomes
the lasso regularization}

\item{weights}{samples weights}

\item{penalty.factor}{weights for the elatic net regularization}

\item{zeroSumWeights}{weights of the zeroSum constraint}

\item{cSum}{constant c of the zeroSum constraint}

\item{standardize}{standardize x and y}

\item{gamma}{penalizing parameter of the fusion term}

\item{fusion}{penalizing matrix of the fusion term}

\item{useOffset}{determines if an offset should be used in the
model or not (TRUE/FALSE)}

\item{useApprox}{determines if the quadratic approximation of the
log-likelihood or the log-likelihood itself should be used
by the local search algorithm for fitting binomial or
multinomial models}

\item{downScaler}{allows to reduce the number of moves}

\item{type}{choose the regression type:
\describe{
        \item{gaussian:}{}
        \item{gaussianZS:}{}
        \item{fusionGaussian:}{}
        \item{fusionGaussianZS:}{}
        \item{binomial:}{}
        \item{binomialZS:}{}
        \item{fusionBinomial:}{}
        \item{fusionBinomialZS:}{}
        \item{multinomial:}{}
        \item{multinomialZS:}{}
        \item{fusionMultinomial:}{}
        \item{fusionMultinomialZS:}{}
}}

\item{algorithm}{choose an algorithm:
\describe{
\item{CD:}{ Coordinate descent (very fast, not so accurate)}
\item{CD+LS:}{ Coordinate descent + Local search (fast, very accurate)}
\item{LS:}{ Local search (slow, accurate)} }}

\item{precision}{stopping criterion of the used algorithms. Determines how
small the improvement of the cost function has to be to stop
the algorithm. Default is 1e-6.}

\item{diagonalMoves}{allows the coordinate descent to use diagonal moves}

\item{polish}{enables a local search at the end of CD to polish the result}

\item{verbose}{verbose = TRUE enables output}

\item{beta}{start coeffiencts of algorithm}
}
\value{
zeroSumFitObject
}
\description{
Creates a linear model with the elastic-net regularization and the zero-sum
     constraint for given x, y, lambda and alpha.
}
\examples{
set.seed(1)
x <- log2(exampleData$x+1)
y <- exampleData$y
zeroSumFit( x, y, lambda=1.5 )

}
